+++ web/src/beta/features/Editor/Map/LayerStylePanel/Editor/NodeSystem/Polygon/polygonNodes/FillColorNode.tsx
