+++ web/src/beta/features/Editor/Map/LayerStylePanel/Editor/NodeSystem/Polyline/polylineNodes/index.tsx
